+++ non_games/procyon/procyon_ai/config/ai_computer_vision_openvino_cpu.def
