+++ non_games/procyon/procyon_ai/config/ai_computer_vision_winml_gpu.def
