+++ non_games/procyon/procyon_ai_img_gen/config/ai_imagegeneration_sd15fp16_openvino.def
