+++ non_games/procyon/procyon_ai_img_gen/config/ai_imagegeneration_sd15int8_openvino.def
