+++ non_games/procyon/procyon_ai_img_gen/config/ai_imagegeneration_sdxlfp16_tensorrt.def
