+++ non_games/procyon/procyon_ai_img_gen/procyon_ai_img_gen_utils.py
