+++ non_games/procyon/procyon_ai_text_generation/config/ai_textgeneration_all.def
